+++ src/utils/Point.java
@@ -3,26 +3,26 @@ package utils;
 import java.io.Serializable;
 
 public class Point implements Serializable {
-	int x;
-	int y;
-	public Point(int x, int y) {
+	float x;
+	float y;
+	public Point(float x, float y) {
 		this.x=x;
 		this.y=y;
 	}
-	public int getX() {
+	public float getX() {
 		return x;
 	}
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 	public void move(int x,int y){
 		this.x=x;
 		this.y=y;
 	}
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
 }
